# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowIQ - AI-powered workflow management platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /companies:
    get:
      operationId: listCompanies
      tags: [companies]
      summary: List all companies (super admin only)
      responses:
        "200":
          description: List of companies
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Company"
    post:
      operationId: createCompany
      tags: [companies]
      summary: Register a new company
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
  /companies/{companyId}:
    get:
      operationId: getCompany
      tags: [companies]
      summary: Get company details
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Company
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"
    patch:
      operationId: updateCompany
      tags: [companies]
      summary: Update company details
      parameters:
        - name: companyId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CompanyUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Company"

  # ─── DEPARTMENTS ─────────────────────────────────────────────────────────────
  /departments:
    get:
      operationId: listDepartments
      tags: [departments]
      summary: List departments for current company
      responses:
        "200":
          description: List of departments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Department"
    post:
      operationId: createDepartment
      tags: [departments]
      summary: Create a new department
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
  /departments/{departmentId}:
    get:
      operationId: getDepartment
      tags: [departments]
      summary: Get department details
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Department
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    patch:
      operationId: updateDepartment
      tags: [departments]
      summary: Update department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepartmentUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Department"
    delete:
      operationId: deleteDepartment
      tags: [departments]
      summary: Delete department
      parameters:
        - name: departmentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # ─── USERS ───────────────────────────────────────────────────────────────────
...
  /requests/{requestId}/comments:
    get:
      operationId: listComments
      tags: [comments]
      summary: List comments on a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Comment"
    post:
      operationId: createComment
      tags: [comments]
      summary: Add comment to a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Comment"

  # ─── NOTIFICATIONS ────────────────────────────────────────────────────────────
...
  /ai/predict/{requestId}:
    get:
      operationId: predictRequestDelay
      tags: [ai]
      summary: Predict delay probability for a request
      parameters:
        - name: requestId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Delay prediction
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DelayPrediction"
  /ai/chat:
    post:
      operationId: chatWithAi
      tags: [ai]
      summary: Chat with the AI executive assistant
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiChatInput"
      responses:
        "200":
          description: AI response
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiChatResponse"
...
  /ai/optimize/{workflowId}:
    get:
      operationId: optimizeWorkflow
      tags: [ai]
      summary: Get AI workflow optimization suggestions
      parameters:
        - name: workflowId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Optimization suggestions
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/WorkflowOptimization"
  /ai/simulate:
    post:
      operationId: runWhatIfSimulation
      tags: [ai]
      summary: Run AI what-if simulation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SimulationInput"
      responses:
        "200":
          description: Simulation result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SimulationResult"
  /ai/classify:
    post:
      operationId: classifyRequest
      tags: [ai]
      summary: AI auto-classify a request description
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClassifyInput"
      responses:
        "200":
          description: Classification result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClassifyResult"
components:
  schemas:
...
    Company:
      type: object
      required: [id, name, slug, plan, employeeCount, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        logo:
          type: ["string", "null"]
        industry:
          type: ["string", "null"]
        plan:
          type: string
          enum: [starter, growth, enterprise]
        employeeCount:
          type: integer
        createdAt:
          type: string
    CompanyInput:
      type: object
      required: [name, slug, adminEmail, adminName]
      properties:
        name:
          type: string
        slug:
          type: string
        industry:
          type: string
        adminEmail:
          type: string
        adminName:
          type: string
        logo:
          type: string
    CompanyUpdate:
      type: object
      properties:
        name:
          type: string
        industry:
          type: string
        logo:
          type: string
        plan:
          type: string

    # ─── DEPARTMENT ─────────────────────────────────────────────────────────────
    Department:
      type: object
      required: [id, companyId, name, memberCount, avgCompletionHours, slaComplianceRate]
      properties:
        id:
          type: integer
        companyId:
          type: integer
        name:
          type: string
        description:
          type: ["string", "null"]
        managerId:
          type: ["integer", "null"]
        managerName:
          type: ["string", "null"]
        memberCount:
          type: integer
        avgCompletionHours:
          type: number
        slaComplianceRate:
          type: number
        activeRequests:
          type: integer
        color:
          type: ["string", "null"]
    DepartmentInput:
      type: object
      required: [name]
      properties:
        name:
          type: string
        description:
          type: string
        managerId:
          type: integer
        color:
          type: string
    DepartmentUpdate:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        managerId:
          type: integer
        color:
          type: string

    # ─── USER ───────────────────────────────────────────────────────────────────
...
    Comment:
      type: object
      required: [id, requestId, content, authorId, authorName, createdAt]
      properties:
        id:
          type: integer
        requestId:
          type: integer
        content:
          type: string
        authorId:
          type: integer
        authorName:
          type: string
        authorAvatar:
          type: ["string", "null"]
        createdAt:
          type: string
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string

    # ─── NOTIFICATION ─────────────────────────────────────────────────────────────
...
    DelayPrediction:
      type: object
      required: [requestId, delayProbability, estimatedCompletionDate,
                 estimatedRemainingDays, confidenceScore]
      properties:
        requestId:
          type: integer
        delayProbability:
          type: number
        estimatedCompletionDate:
          type: string
        estimatedRemainingDays:
          type: number
        riskLevel:
          type: string
          enum: [low, medium, high, critical]
        confidenceScore:
          type: number
        factors:
          type: array
          items:
            type: string
    AiChatInput:
      type: object
      required: [message]
      properties:
        message:
          type: string
        conversationId:
          type: string
    AiChatResponse:
      type: object
      required: [message, conversationId]
      properties:
        message:
          type: string
        conversationId:
          type: string
        relatedRequestIds:
          type: array
          items:
            type: integer
        dataReferences:
          type: array
          items:
            type: string
...
    WorkflowOptimization:
      type: object
      required: [workflowId, currentAvgHours, optimizedAvgHours, improvementPercent, suggestions]
      properties:
        workflowId:
          type: integer
        currentAvgHours:
          type: number
        optimizedAvgHours:
          type: number
        improvementPercent:
          type: number
        suggestions:
          type: array
          items:
            $ref: "#/components/schemas/OptimizationSuggestion"
    OptimizationSuggestion:
      type: object
      required: [type, description, estimatedImprovementPercent, confidence]
      properties:
        type:
          type: string
          enum: [remove_step, reorder_steps, merge_approvals, add_resources, change_assignee]
        description:
          type: string
        estimatedImprovementPercent:
          type: number
        confidence:
          type: number
    SimulationInput:
      type: object
      required: [scenario]
      properties:
        scenario:
          type: string
        departmentId:
          type: integer
        additionalEmployees:
          type: integer
        workflowId:
          type: integer
    SimulationResult:
      type: object
      required: [scenario, currentMetrics, simulatedMetrics, improvements]
      properties:
        scenario:
          type: string
        currentMetrics:
          $ref: "#/components/schemas/MetricsSnapshot"
        simulatedMetrics:
          $ref: "#/components/schemas/MetricsSnapshot"
        improvements:
          type: array
          items:
            $ref: "#/components/schemas/ImprovementItem"
    MetricsSnapshot:
      type: object
      required: [avgCompletionHours, slaComplianceRate, bottleneckCount]
      properties:
        avgCompletionHours:
          type: number
        slaComplianceRate:
          type: number
        bottleneckCount:
          type: integer
        delayedRequestsPercent:
          type: number
    ImprovementItem:
      type: object
      required: [metric, currentValue, simulatedValue, changePercent]
      properties:
        metric:
          type: string
        currentValue:
          type: number
        simulatedValue:
          type: number
        changePercent:
          type: number
    ClassifyInput:
      type: object
      required: [description]
      properties:
        description:
          type: string
        title:
          type: string
    ClassifyResult:
      type: object
      required: [category, priority, expectedCompletionHours, confidence]
      properties:
        category:
          type: string
        priority:
          type: string
        expectedCompletionHours:
          type: number
        requiredDepartments:
          type: array
          items:
            type: string
        confidence:
          type: number
        suggestedWorkflowId:
          type: ["integer", "null"]